Refactor #4 - Avoid repeating names on the imports
Using, index file, other files located in another folder can import them at once.

This resolves the redunduant import statments.
