+++ src/components/Tasks/index.js
@@ -1,0 +1,5 @@
+import Task from './Task';
+import TaskForm from './TaskForm';
+import TaskList from './TaskList';
+
+export { Task, TaskForm, TaskList };
